requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai==0.7.2
tensorflow==2.16.1
Pillow==10.3.0
SpeechRecognition==3.10.4
pydub==0.25.1
numpy==1.26.4
